+++ poolstats.tcl
@@ -1,12 +1,38 @@
-# Pool Stats
+#
+# MPOS eggdrop Calls
+#
 
-set scriptversion "v0.1 ALPHA"
-set help_blocktime "10"
+set scriptversion "v0.2"
 
-set apiurl "http://yourdomain.td/"
+# time to wait before next command in seconds
+#
+set help_blocktime "5"
+
+# interval to check for new blocks in seconds
+# if set to 0, the bot will do no automatic
+# check for new blocks
+# checktime is seconds
+#
+set blockchecktime "0"
+
+# channels to advertise new blocks
+# and run commands
+#
+set channels "#firstchannel #secondchannel #thirdchannel"
+
+# url where mpos is installed
+#
+set apiurl "http://yourdomain.tld/"
+
+# api key from mpos
+#
 set apikey "YOURAPIKEYFROMMPOS"
 
 
+
+
+# Key Bindings
+#
 bind pub - !pool pool_info
 bind pub - !block block_info
 bind pub - !last last_info
@@ -15,6 +41,24 @@ bind pub - !round round_info
 bind pub - !help printUsage
 
 
+
+
+# start timer if set
+# and check if started when bot makes a rehash
+#
+if {$blockchecktime ne "0"} {
+	# check if timer is running
+	# else a rehash starts a new timer
+	if {![info exists checknewblocks_running]} {
+		putlog "Timer aktiviert"
+ 	   	utimer $blockchecktime checknewblocks
+	    set checknewblocks_running 1
+	}
+}
+
+
+                                  
+  
 
 proc printUsage {nick host hand chan arg} {
     putquick "NOTICE $nick :Usage: !block       - Blockstats"
@@ -25,8 +69,82 @@ proc printUsage {nick host hand chan arg} {
     putquick "NOTICE $nick :       !help        - This help text"
 }
 
+	
+
+                                         
+
+proc checknewblocks {} {
+	global blockchecktime channels apiurl apikey channels
+	package require http
+	package require json
+	
+ 	set action "index.php?page=api&action=getblocksfound&limit=1&api_key="
+ 	
+  	set newurl $apiurl
+  	append newurl $action
+  	append newurl $apikey
+  	
+    set token [::http::geturl "$newurl"]
+    set data [::http::data $token]
+    ::http::cleanup $token
+    #putlog "xml: $data"
+    set results [::json::json2dict $data]
+	
+	foreach {key value} $results {
+		#putlog "Key: $key - $value"
+		foreach {sub_key sub_value} $value {
+			#putlog "Sub: $sub_key - $sub_value"
+			if {$sub_key eq "data"} {
+				#putlog "Sub: $sub_value"
+				foreach {elem elem_val} $sub_value {
+					#putlog "Ele: $elem - Val: $elem_val"
+					foreach {elem2 elem_val2} $elem {
+						#putlog "Ele: $elem2 - Val: $elem_val2"
+
+      					if {$elem2 eq "time"} {
+      						# used actual timestamp - 2x $blockchecktime + 1
+      						# it's used for checking if block is found a second
+      						# after last check and api is cached or memcached
+      						set time1 [expr [unixtime]-$blockchecktime-$blockchecktime+1]
+      						set time2 $elem_val2
+      					}
+      					if {$elem2 eq "height"} { set last_block "Last Block: $elem_val2" }
+      					if {$elem2 eq "shares"} { set last_shares "Shares: $elem_val2" } 
+						if {$elem2 eq "finder"} { set last_finder "Finder: $elem_val2" }
+						if {$elem2 eq "confirmations"} { 
+							if {$elem_val2 eq "-1"} { 
+								set last_status "Status: Orphaned"
+							} else {
+								set last_status "Status: Valid"
+							}
+						}
+						
+					}
+					break
+				}
+			}
+		}
+	}
+	
+	
+	if {$time1 < $time2} {
+		foreach advert $channels {
+			putquick "PRIVMSG $advert :New Block Found"
+ 			putquick "PRIVMSG $advert :$last_block | $last_status | $last_shares | $last_finder"
+ 		}
+	} else {
+		set timelast [strftime "%d.%m.%Y - %T" $time2]
+		set timetocheck [strftime "%d.%m.%Y - %T" $time1]
+		putlog "No New Block: lastfind: $timelast | checktime: $timetocheck"
+	}
+	
+	utimer $blockchecktime checknewblocks
+
+  }                                      
+  
+  
 proc user_info {nick host hand chan arg} {
- 	global apiurl apikey help_blocktime help_blocked
+ 	global apiurl apikey help_blocktime help_blocked channels
  	
  	set action "index.php?page=api&action=getuserstatus&id=$arg&api_key="
  	
@@ -73,8 +191,10 @@ proc user_info {nick host hand chan arg} {
 		}
 	}
 	
-	putquick "PRIVMSG $chan :User Info for $arg"
-	putquick "PRIVMSG $chan :$user_hashrate | $user_validround | $user_invalidround | $user_sharerate"
+	foreach advert $channels {
+		putquick "PRIVMSG $advert :User Info for $arg"
+		putquick "PRIVMSG $advert :$user_hashrate | $user_validround | $user_invalidround | $user_sharerate"
+	}
 
 	set help_blocked($mask) 1
 	utimer $help_blocktime [ list unset help_blocked($mask) ]
@@ -84,7 +204,7 @@ proc user_info {nick host hand chan arg} {
 
 
 proc round_info {nick host hand chan arg } {
- 	global apiurl apikey help_blocktime help_blocked
+ 	global apiurl apikey help_blocktime help_blocked channels
 	package require http
 	package require json
 	
@@ -137,10 +257,12 @@ proc round_info {nick host hand chan arg } {
 	}
 	
 	set allshares [expr $shares_valid+$shares_invalid]
-	
-	putquick "PRIVMSG $chan :Actual Round"
- 	putquick "PRIVMSG $chan :$shares_estimated | Sharecount: $allshares | Shares valid: $shares_valid | Shares invalid: $shares_invalid | $shares_progress"
 
+	foreach advert $channels {
+		putquick "PRIVMSG $advert :Actual Round"
+ 		putquick "PRIVMSG $advert :$shares_estimated | Sharecount: $allshares | Shares valid: $shares_valid | Shares invalid: $shares_invalid | $shares_progress"
+	}
+	
 	set help_blocked($mask) 1
 	utimer $help_blocktime [ list unset help_blocked($mask) ]
 
@@ -152,7 +274,7 @@ proc round_info {nick host hand chan arg } {
 
 
 proc last_info {nick host hand chan arg } {
- 	global apiurl apikey help_blocktime help_blocked
+ 	global apiurl apikey help_blocktime help_blocked channels
 	package require http
 	package require json
 	
@@ -206,9 +328,11 @@ proc last_info {nick host hand chan arg } {
 		}
 	}
 	
-	putquick "PRIVMSG $chan :Last Block"
- 	putquick "PRIVMSG $chan :$last_block | $last_confirmed | $last_difficulty | $last_timefound | $last_shares | $last_estshares | $last_finder"
-
+	foreach advert $channels {
+		putquick "PRIVMSG $advert :Last Block"
+ 		putquick "PRIVMSG $advert :$last_block | $last_confirmed | $last_difficulty | $last_timefound | $last_shares | $last_estshares | $last_finder"
+ 	}
+ 	
 	set help_blocked($mask) 1
 	utimer $help_blocktime [ list unset help_blocked($mask) ]
 
@@ -218,7 +342,7 @@ proc last_info {nick host hand chan arg } {
 
 # Pool Stats
 proc pool_info {nick host hand chan arg} {
-    global apiurl apikey help_blocktime help_blocked
+    global apiurl apikey help_blocktime help_blocked channels
 	package require http
 	package require json
 	
@@ -260,17 +384,19 @@ proc pool_info {nick host hand chan arg} {
 			}
 		}
 	}
-
-    putquick "PRIVMSG $chan :Pool Stats"
-    putquick "PRIVMSG $chan :$pool_hashrate | $pool_efficiency | $pool_workers | $pool_nethashrate"
-
+	
+	foreach advert $channels {
+    	putquick "PRIVMSG $advert :Pool Stats"
+    	putquick "PRIVMSG $advert :$pool_hashrate | $pool_efficiency | $pool_workers | $pool_nethashrate"
+    }
+    
 }
 
 
 
 # Block Stats
 proc block_info {nick host hand chan arg} {
-    global apiurl apikey help_blocktime help_blocked
+    global apiurl apikey help_blocktime help_blocked channels
 	package require http
 	package require json
 	
@@ -325,17 +451,13 @@ proc block_info {nick host hand chan arg} {
 			}
 		}
 	}
-
-    putquick "PRIVMSG $chan :Block Stats"
-    putquick "PRIVMSG $chan :$block_current | $block_next | $block_last | $block_diff | $block_time | $block_shares | $block_timelast | $block_last"
-
+	
+	foreach advert $channels {
+    	putquick "PRIVMSG $chan :Block Stats"
+    	putquick "PRIVMSG $chan :$block_current | $block_next | $block_last | $block_diff | $block_time | $block_shares | $block_timelast | $block_last"
+    }
+    
 }
 
 
-
-
-
-
-
-
-putlog "===>> Mining-Pool-Stats - Version $scriptversion"
+putlog "===>> Mining-Pool-Stats - Version $scriptversion loaded"