+++ Changelog.md
@@ -1,3 +1,10 @@
+v. 0.2
+
+* added some more settings
+  - channels the bot sits in and should post news
+  - blockchecktime (used to check for new blocks)
+* added blockchecking and announcing new block to channel
+
 v. 0.1 ALPHA
 
 * initial release